Ceph OSD disk targeting
This allows users to configure ceph such that an OSD pod will
be created fror each device that is provided.

OSD devices and journals may be specified either using paths to
block device files or with SCSI paths of the form "scsi@x:y:z:n".
Journal devices may also have a partition specified either as
part of the device name or as a separate value.

The patch also updates liveness and readiness checks to be more
appropriate for an environment with multiple OSDs per host.

Change-Id: Ibbb535a8753b4bada49e4b0139489f5966854b5c
Signed-off-by: Ganesh Maharaj Mahalingam <ganesh.mahalingam@intel.com>